+++ R/pwrA_S3methods.R
@@ -1,3 +1,23 @@
+# --------------------------------------------------------------------------
+# shadowtext() lifted from the package TeachingDemos
+# author Greg Snow, orphaned on CRAN 2024-02-10
+# request from Prof Brian Ripley to comply with the CRAN policy
+shadowtext <- function(x, y = NULL, labels, col = "white", bg = "black",
+                       theta = seq(pi / 32, 2 * pi, length.out = 64),
+                       r = 0.1, cex = 1, ... ) {
+  xy  <- xy.coords(x, y)
+  fx  <- grconvertX(xy$x, to = "nfc")
+  fy  <- grconvertY(xy$y, to = "nfc")
+  fxo <- r * strwidth("A", units = "figure", cex = cex)
+  fyo <- r * strheight("A", units = "figure", cex = cex)
+  for (i in theta) {
+    text(grconvertX(fx + cos(i) * fxo, from = "nfc"),
+         grconvertY(fy + sin(i) * fyo, from = "nfc"),
+         labels, cex = cex, col = bg, ...)
+  }
+  text(xy$x, xy$y, labels, cex = cex, col = col, ... )
+}
+
 # S3 method for printing the results of pa.ABE(), pa.scABE()
 # --------------------------------------------------------------------------
 print.pwrA <- function(x, digits=4, plotit=TRUE, ...)